step_aware_preference_model/requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@d1aa558119859c4b205a324afabaecabd9ef375e
datasets==2.10.1
deepspeed==0.12.3
fire==0.4.0
hydra-core==1.3.2
rich==13.3.2
submitit==1.4.5
transformers==4.27.3
wandb==0.12.21